Added support for overriding arbitrary packaging props
By turning the @(NuGetize) custom metadata into AdditionalProperties we can
pass them when building the package and allow the item definition to control
all properties of a package, including the target package id, project URL and
so on.

Added support for selecting either Mono versioning or Assembly versioning for
the package. For the later case, we also add token replacement of $MonoVersion$
in the Description and PackageReleaseNotes so that you can provide a friendly
description that mentions what Mono the given assembly first shipped with.